this works.
I already made some optimizations (cache the index) and such, but with
part 2 things get *way* bigger and slower and those array operations are
probably horribly expensive. It's likely I'm going to need to convert
this into a linked list instead of using an array.

honestly, this has been running for like 10 minutes already, and I'm
still fairly confident I could convert this whole mess to a linked list
and run it before this finishes