ok, after messing around with the program more I figured out a few things
the 17 -> 31 "function" sets registers 4 and 5. register 4 is a red
herring. it's immediately overwritten inside the loop.

register 5 is our actual "input" to the rest of the program. We'll call
it N

the rest of the program is a loop:
output = 0
for i in 1..N
	for j in i..N
		if i*j == N
			output += i # or j, can't remember. it's irrelevant

what is this doing?

it's looking for pairs of numbers that multiply to produce N. and
summing them

And it's not doing it very efficiently. At all. But that's ok.

Because now that we have what it's doing, we can just plug that number
into a factor pair calculator (like
https://www.calculatorsoup.com/calculators/math/factors.php, which is
what I used) sum those, and get the answer we are looking for.

This was challenging, and quite fun. yay